# internal/api/openapi.yaml
openapi: "3.0.0"
info:
  contact:
    email: n@nickysemenza.com
  description: API for https://github.com/nickysemenza/gourd
  version: 1.0.0
  title: Gourd Recipe Database
  license:
    name: MIT
    # identifier: MIT
...
servers:
  - description: test
    url: http://localhost:4242/api
  - description: prod
    url: https://food.nickysemenza.com/api
paths:
...
  /recipes/scrape:
    post:
      summary: scrape a recipe by URL
      description: todo
      operationId: scrape_recipe
      tags:
        - recipes
      requestBody:
        description: sum stuff
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  type: string
      responses:
        "201":
          description: Newly created recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeWrapper"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
security:
  - bearerAuth: [] # use the same name as above
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for documentation purposes
  schemas:
    SectionIngredient:
      description: Ingredients in a single section
      type: object
      required:
        - id
        - amounts
        - kind
      properties:
        id:
          description: id
          type: string
        kind:
          description: what kind of ingredient
          $ref: "#/components/schemas/IngredientKind"
        recipe:
          $ref: "#/components/schemas/RecipeDetail"
        ingredient:
          $ref: "#/components/schemas/IngredientDetail"
        amounts:
          description: the various measures
          type: array
          items:
            $ref: "#/components/schemas/Amount"
        adjective:
          description: adjective
          type: string
        optional:
          description: optional
          type: boolean
        original:
          description: raw line item (pre-import/scrape)
          type: string
        substitutes:
          description: x
          type: array
          items:
            $ref: "#/components/schemas/SectionIngredient"
...
    SectionInstruction:
      description: Instructions in a single section
      type: object
      required:
        - id
        - instruction
      properties:
        id:
          description: id
          type: string
        instruction:
          description: instruction
          type: string
...
    RecipeSection:
      description: A step in the recipe
      type: object
      required:
        - id
        - instructions
        - ingredients
      properties:
        id:
          description: id
          type: string
        duration:
          $ref: "#/components/schemas/Amount"
        instructions:
          description: x
          type: array
          items:
            $ref: "#/components/schemas/SectionInstruction"
        ingredients:
          description: x
          type: array
          items:
            $ref: "#/components/schemas/SectionIngredient"
...
    RecipeWrapper:
      description: A recipe with subcomponents
      type: object
      required:
        - detail
        - id
      properties:
        id:
          description: id
          type: string
        detail:
          $ref: "#/components/schemas/RecipeDetail"
        linked_meals:
          type: array
          items:
            $ref: "#/components/schemas/Meal"
        linked_photos:
          type: array
          items:
            $ref: "#/components/schemas/Photo"
...
    RecipeDetail:
      description: A revision of a recipe
      type: object
      required:
        - id
        - name
        - sources
        - quantity
        - unit
        - sections
        - version
        - is_latest_version
        - created_at
        - tags
      properties:
        id:
          description: id
          type: string
        sections:
          description: sections of the recipe
          type: array
          items:
            $ref: "#/components/schemas/RecipeSection"
        name:
          description: recipe name
          type: string
        sources:
          description: book or websites
          type: array
          items:
            $ref: "#/components/schemas/RecipeSource"
        servings:
          description: num servings
          type: integer
          format: int64
        quantity:
          description: serving quantity
          type: integer
          format: int64
        unit:
          description: serving unit
          type: string
        version:
          description: version of the recipe
          type: integer
          format: int64
        is_latest_version:
          description: whether or not it is the most recent version
          type: boolean
        created_at:
          description: when the version was created
          type: string
          format: date-time
        other_versions:
          description: Other versions
          type: array
          items:
            $ref: "#/components/schemas/RecipeDetail"
        tags:
          description: tags
          type: array
          items:
            type: string
...
    Ingredient:
      description: "An Ingredient"
      type: object
      required:
        - id
        - name
      properties:
        id:
          description: id
          type: string
        name:
          description: Ingredient name
          type: string
        parent:
          description: ingredient ID for a similar (likely a different spelling)
          type: string
        fdc_id:
          description: FDC id equivalent to this ingredient
          type: integer
          format: int64
    IngredientDetail:
      description: "An Ingredient"
      type: object
      required:
        - ingredient
        - recipes
        # - children
        - unit_mappings
      properties:
        ingredient:
          $ref: "#/components/schemas/Ingredient"
        recipes:
          description: Recipes referencing this ingredient
          type: array
          items:
            $ref: "#/components/schemas/RecipeDetail"
        children:
          description: Ingredients that are equivalent
          type: array
          items:
            $ref: "#/components/schemas/IngredientDetail"
        food:
          $ref: "#/components/schemas/Food"
        unit_mappings:
          description: mappings of equivalent units
          type: array
          items:
            $ref: "#/components/schemas/UnitMapping"
    UnitMapping:
      description: mappings
      type: object
      required:
        - a
        - b
      properties:
        a:
          $ref: "#/components/schemas/Amount"
        b:
          $ref: "#/components/schemas/Amount"
        source:
          description: source of the mapping
          type: string
...
    Amount:
      description: amount and unit
      type: object
      required:
        - unit
        - value
      properties:
        unit:
          description: unit
          type: string
        value:
          description: value
          type: number
          format: double # float64
        upper_value:
          description: value
          type: number
          format: double # float64
        source:
          description: if it was explicit, inferred, etc
          type: string
    RecipeSource:
      description: where the recipe came from (i.e. book/website)
      type: object
      properties:
        url:
          description: url
          type: string
        title:
          description: title (if book)
          type: string
        page:
          description: page number/section (if book)
          type: string
        image_url:
          description: image url
          type: string
...
    Photo:
      description: A photo
      type: object
      required:
        - id
        - base_url
        - created
        - width
        - height
        - source
      properties:
        id:
          description: id
          type: string
        base_url:
          description: public image
          type: string
        blur_hash:
          description: blur hash
          type: string
        created:
          description: when it was taken
          type: string
          format: date-time
        width:
          description: width px
          type: integer
          format: int64
        height:
          description: height px
          type: integer
          format: int64
        source:
          description: where the photo came from
          type: string
          enum: [google, notion]
    Meal:
      description: A meal, which bridges recipes to photos
      type: object
      required:
        - id
        - name
        - ate_at
        - photos
      properties:
        id:
          description: id
          type: string
        name:
          description: public image
          type: string
        ate_at:
          description: when it was taken
          type: string
          format: date-time
        photos:
          type: array
          items:
            $ref: "#/components/schemas/Photo"
        recipes:
          type: array
          items:
            $ref: "#/components/schemas/MealRecipe"
    MealRecipe:
      description: A recipe that's part of a meal (a recipe at a specific amount)
      type: object
      required:
        - multiplier
        - recipe
      properties:
        multiplier:
          description: when it was taken
          type: number
          format: double # float64
        recipe:
          $ref: "#/components/schemas/RecipeDetail"
...
    Error:
      type: object
      description: A generic error message
      required: ["message"]
      properties:
        message:
          type: string
        trace_id:
          type: string
      example:
        message: "Something went wrong"
...
    Nutrient:
      type: object
      description: todo
      required:
        - id
        - name
        - unit_name
      properties:
        id:
          description: todo
          type: integer
        name:
          description: todo
          type: string
        unit_name:
          description: todo
          $ref: "#/components/schemas/FoodNutrientUnit"
    FoodNutrient:
      type: object
      description: todo
      required:
        - nutrient
        - amount
        - data_points
      properties:
        nutrient:
          $ref: "#/components/schemas/Nutrient"
        amount:
          type: number
          format: double # float64
        data_points:
          type: integer
    FoodCategory:
      type: object
      description: food category, set for some
      required:
        - code
        - description
      properties:
        code:
          description: Food description
          type: string
        description:
          description: Food description
          type: string
    BrandedFood:
      type: object
      description: branded_food
      required:
        - serving_size
        - serving_size_unit
      properties:
        brand_owner:
          type: string
        ingredients:
          type: string
        serving_size:
          type: number
          format: double # float64
        serving_size_unit:
          type: string
        household_serving:
          type: string
        branded_food_category:
          type: string
    FoodPortion:
      type: object
      description: food_portion
      required:
        - id
        - amount
        - portion_description
        - modifier
        - gram_weight
      properties:
        id:
          type: integer
        amount:
          type: number
          format: double # float64
        portion_description:
          type: string
        modifier:
          type: string
        gram_weight:
          type: number
          format: double # float64
    Food:
      type: object
      description: A top level food
      required:
        - fdc_id
        - description
        - data_type
        - nutrients
        - unit_mappings
      properties:
        fdc_id:
          description: FDC Id
          type: integer
        description:
          description: Food description
          type: string
        data_type:
          description: todo
          $ref: "#/components/schemas/FoodDataType"
        category:
          $ref: "#/components/schemas/FoodCategory"
        nutrients:
          description: todo
          type: array
          items:
            $ref: "#/components/schemas/FoodNutrient"
        portions:
          description: portion datapoints
          type: array
          items:
            $ref: "#/components/schemas/FoodPortion"
        branded_info:
          $ref: "#/components/schemas/BrandedFood"
        unit_mappings:
          description: mappings of equivalent units
          type: array
          items:
            $ref: "#/components/schemas/UnitMapping"
...
    FoodDataType:
      type: string
      enum:
        [
          foundation_food,
          sample_food,
          market_acquisition,
          survey_fndds_food,
          sub_sample_food,
          agricultural_acquisition,
          sr_legacy_food,
          branded_food,
        ]
    FoodNutrientUnit:
      type: string
      enum: [UG, G, IU, kJ, KCAL, MG, MG_ATE, SP_GR]
    IngredientKind:
      type: string
      enum: [ingredient, recipe]
...